dev.requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-mypy==17.8.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
pep8-naming==0.13.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
testfixtures==6.18.5
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
